+++ lib/i18next.dart
@@ -2,4 +2,5 @@ library i18next;
 
 export 'src/i18next.dart';
 export 'src/options.dart';
+export 'src/resource_store.dart';
 export 'src/utils.dart';