+++ lib/src/utils.dart
@@ -1,5 +1,3 @@
 import 'dart:ui';
 
 typedef ArgumentFormatter = String Function(Object, String, Locale);
-
-typedef LocalizationDataSource = Map<String, Object> Function(String, Locale);